pyproject.toml
[project]
name = "summariser"
version = "0.1.0"
description = ""
packages = [{include = "summariser"}]
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "hydra-core (>=1.3.2,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "flask (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flask-cors (==4.0.1)",
    "loguru (==0.7.2)",
    "ratelimit (>=2.2.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
